Refactored unit tests for websocket sessions
Extracted most of the reusable classes and methods from `WebsocketEndpointTest` into a new `AbstractWebsocketTestBase` class that can be reused in other test classes.

This makes no changes to the actual test methods in `WebsocketEndpointTest`, and does not add or remove tests.
